% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big.factor.R
\name{as.big.factor}
\alias{as.big.factor}
\title{Coerce a vector of character to a big.factor, quick-and-dirty for testing,
no frills, etc...}
\usage{
as.big.factor(x)
}
\arguments{
\item{x}{a vector of character (for now)}
}
\value{
a \code{\link{big.factor}}
}
\description{
Coerce a vector of character to a big.factor, quick-and-dirty for testing,
no frills, etc...
}
