% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{is.big.char,big.char-method}
\alias{[,big.char,ANY,missing,missing-method}
\alias{[,big.char,missing,missing,missing-method}
\alias{[<-,big.char,ANY,missing-method}
\alias{[<-,big.char,missing,missing-method}
\alias{is.big.char,ANY-method}
\alias{is.big.char,big.char-method}
\alias{length,big.char-method}
\alias{names,big.char-method}
\alias{names<-,big.char,character-method}
\title{Do we have a big.char?}
\usage{
\S4method{is.big.char}{big.char}(x)

\S4method{is.big.char}{ANY}(x)

\S4method{length}{big.char}(x)

\S4method{names}{big.char}(x)

\S4method{names}{big.char,character}(x) <- value

\S4method{[}{big.char,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.char,ANY,missing}(x, i, j, ...) <- value

\S4method{[}{big.char,missing,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.char,missing,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a \code{\link{big.char}}}

\item{value}{the returned object}

\item{i}{indices (or equivalent) for extraction}

\item{j}{typically not used or supported}

\item{...}{doesn't currently have a role, but may for \code{stringsAsFactors}}

\item{drop}{what you would expect when the returned object is of length 1}
}
\description{
Do we have a big.char?

Get the length of a big.char vector

Get the names of a big.char vector

Set the names of a big.char vector

Core big.char extraction

Core big.char assignment

Full big.char extraction

Full big.char assignment
}

