% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big.char-package.R
\docType{package}
\name{big.char-package}
\alias{big.char-package}
\title{Provides very large vectors of character strings}
\description{
This package extends bigmemory's big.matrix objects
to support big vectors of character strings.  The only constraint
is the user-specified maximum character length which must be
satisfied for every string in the vector.  Shorter strings are
padded with NAs, consuming (and wasting) space, but this is
unavoidable given the use of memory-mapped files.
}
\author{
Jay Emerson
}
