% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{big.factor-class}
\alias{big.factor-class}
\title{S4 class big.factor contains a big.matrix (vector) of integers
and a big.char of factor levels.
exportClass big.factor}
\description{
S4 class big.factor contains a big.matrix (vector) of integers
and a big.char of factor levels.
exportClass big.factor
}

