% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big.factor.R
\name{big.factor}
\alias{big.factor}
\title{Create a big.factor object}
\usage{
big.factor(length, maxchar = 8, init = NULL, levels = NULL,
  backingfile = NULL, backingpath = NULL, descriptorfile = NULL,
  binarydescriptor = FALSE, shared = TRUE)
}
\arguments{
\item{length}{the vector length}

\item{maxchar}{the maximum length of the strings, 8 by default}

\item{init}{an optional string for initialization purposes}

\item{levels}{a vector of character factor levels (labels), which may
be truncated to \code{maxchar} length if necessary.}

\item{backingfile}{optional name of binary memory-mapped file}

\item{backingpath}{should be obvious, right?}

\item{descriptorfile}{the descriptor file associated with the backingfile}

\item{binarydescriptor}{see \code{\link[bigmemory]{big.matrix}}}

\item{shared}{see \code{\link[bigmemory]{big.matrix}}}
}
\value{
a \code{big.factor}
}
\description{
Create a \code{big.factor} object that could be
in shared memory or larger-than-RAM via \code{\link{big.char}}
and \code{\link[bigmemory]{big.matrix}} objects.
}
\details{
There are real hazards here.  I may not finish it now.
}
\author{
Jay Emerson
}
