% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.big.char}
\alias{is.big.char}
\title{Generic function is.big.char()}
\usage{
is.big.char(x)
}
\arguments{
\item{x}{a \code{\link{big.char}} vector of strings}
}
\description{
Do we have a \code{\link{big.char}}?
}
\details{
No further detail is needed.
}

